Remove consolidated parser test file
