+++ src/parser/tests/round_trip.rs
@@ -1,0 +1,77 @@
+//! Round-trip parser tests.
+//!
+//! These tests ensure parsing followed by pretty-printing reproduces the
+//! original source and that basic program structure is detected correctly.
+
+use super::common::pretty_print;
+use crate::{SyntaxKind, parse};
+use rstest::{fixture, rstest};
+
+#[fixture]
+fn simple_prog() -> &'static str {
+    "input relation R(x: u32);"
+}
+
+#[fixture]
+fn complex_prog() -> &'static str {
+    "input relation R(x: u32);\noutput relation S(y: string);"
+}
+
+#[fixture]
+fn empty_prog() -> &'static str {
+    ""
+}
+
+#[rstest]
+fn parse_round_trip(simple_prog: &str) {
+    let parsed = parse(simple_prog);
+    let text = pretty_print(parsed.root().syntax());
+    assert_eq!(text, simple_prog);
+    assert_eq!(parsed.root().kind(), SyntaxKind::N_DATALOG_PROGRAM);
+}
+
+#[rstest]
+fn complex_program_round_trip(complex_prog: &str) {
+    let parsed = parse(complex_prog);
+    assert!(parsed.errors().is_empty());
+    let text = pretty_print(parsed.root().syntax());
+    assert_eq!(text, complex_prog);
+    let relations = parsed.root().relations();
+    assert_eq!(relations.len(), 2);
+    let [first, second] = relations.as_slice() else {
+        panic!("expected two relations");
+    };
+    assert!(first.is_input());
+    assert!(second.is_output());
+    assert_eq!(first.name(), Some("R".into()));
+    assert_eq!(second.name(), Some("S".into()));
+}
+
+#[rstest]
+fn empty_program_has_no_items(empty_prog: &str) {
+    let parsed = parse(empty_prog);
+    assert!(parsed.errors().is_empty());
+    assert!(parsed.root().imports().is_empty());
+    assert!(parsed.root().type_defs().is_empty());
+    assert!(parsed.root().relations().is_empty());
+    assert!(parsed.root().functions().is_empty());
+    assert!(parsed.root().indexes().is_empty());
+    assert!(parsed.root().rules().is_empty());
+    assert_eq!(pretty_print(parsed.root().syntax()), "");
+    assert_eq!(parsed.root().syntax().children().count(), 0);
+}
+
+#[rstest]
+fn error_token_produces_error_node() {
+    let source = "?";
+    let parsed = parse(source);
+    let root = parsed.root().syntax();
+    let has_error = root
+        .children_with_tokens()
+        .filter_map(|child| match child {
+            rowan::NodeOrToken::Node(node) => Some(node),
+            rowan::NodeOrToken::Token(_) => None,
+        })
+        .any(|node| node.kind() == SyntaxKind::N_ERROR);
+    assert!(has_error);
+}
